// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/TotalCashBalanceDocumentRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-organisation-users
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: @project.version@
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TotalCashBalanceDocumentRequest
 */
export interface TotalCashBalanceDocumentRequest {
    /**
     * 
     * @type {number}
     * @memberof TotalCashBalanceDocumentRequest
     */
    totalCashBalance?: number;
    /**
     * 
     * @type {string}
     * @memberof TotalCashBalanceDocumentRequest
     */
    currency?: string;
    /**
     * 
     * @type {string}
     * @memberof TotalCashBalanceDocumentRequest
     */
    additionalCash?: TotalCashBalanceDocumentRequestAdditionalCashEnum;
    /**
     * 
     * @type {number}
     * @memberof TotalCashBalanceDocumentRequest
     */
    totalAmountDueIn?: number;
}


/**
 * @export
 */
export const TotalCashBalanceDocumentRequestAdditionalCashEnum = {
    WaitingForPaymentsIn2Weeks: 'WAITING_FOR_PAYMENTS_IN_2_WEEKS',
    NotWaitingOnPaymentsIn2Weeks: 'NOT_WAITING_ON_PAYMENTS_IN_2_WEEKS'
} as const;
export type TotalCashBalanceDocumentRequestAdditionalCashEnum = typeof TotalCashBalanceDocumentRequestAdditionalCashEnum[keyof typeof TotalCashBalanceDocumentRequestAdditionalCashEnum];


/**
 * Check if a given object implements the TotalCashBalanceDocumentRequest interface.
 */
export function instanceOfTotalCashBalanceDocumentRequest(value: object): value is TotalCashBalanceDocumentRequest {
    return true;
}

export function TotalCashBalanceDocumentRequestFromJSON(json: any): TotalCashBalanceDocumentRequest {
    return TotalCashBalanceDocumentRequestFromJSONTyped(json, false);
}

export function TotalCashBalanceDocumentRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): TotalCashBalanceDocumentRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'totalCashBalance': json['totalCashBalance'] == null ? undefined : json['totalCashBalance'],
        'currency': json['currency'] == null ? undefined : json['currency'],
        'additionalCash': json['additionalCash'] == null ? undefined : json['additionalCash'],
        'totalAmountDueIn': json['totalAmountDueIn'] == null ? undefined : json['totalAmountDueIn'],
    };
}

export function TotalCashBalanceDocumentRequestToJSON(json: any): TotalCashBalanceDocumentRequest {
    return TotalCashBalanceDocumentRequestToJSONTyped(json, false);
}

export function TotalCashBalanceDocumentRequestToJSONTyped(value?: TotalCashBalanceDocumentRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'totalCashBalance': value['totalCashBalance'],
        'currency': value['currency'],
        'additionalCash': value['additionalCash'],
        'totalAmountDueIn': value['totalAmountDueIn'],
    };
}